Fix for Issue #583
This fix adds the parameter `descendIfEdge`, which changes the behavior of `descendPeak` if set to TRUE. Doing so will continue the descend if the centroid is near either end of the spectrum by trying the full width instead of the half width. Defaults to FALSE to keep the current behavior to preserve current centroid results.